tests/baseline/protocol-1.2/inspector/sources/debugger/properties-special-expected.txt

CONSOLE MESSAGE: line 10: true
CONSOLE MESSAGE: line 11: function (a,b) { return a + b; }
CONSOLE MESSAGE: line 12: function () { [native code] }
CONSOLE MESSAGE: line 13: function* () { yeild [1,2,3] }
Tests how debugger presents special properties of closures, bound functions and object wrappers.

properties-special.html:10 Boolean
    __proto__: Boolean
    [[PrimitiveValue]]: true
properties-special.html:11 function anonymous(a,b)
    arguments: null
    caller: null
    length: 2
    name: ""
    prototype: Object
    __proto__: function ()
    [[FunctionLocation]]: properties-special.html:11
    [[Scopes]]: Scopes[1]
properties-special.html:12 function bound ()
    arguments: (...)
    caller: (...)
    length: 1
    name: "bound "
    __proto__: function ()
    [[TargetFunction]]: function (a,b)
    [[BoundThis]]: Object
    [[BoundArgs]]: Array[1]
properties-special.html:13 function* anonymous()
    arguments: (...)
    caller: (...)
    length: 0
    name: ""
    prototype: Generator
    __proto__: GeneratorFunction
    [[FunctionLocation]]: properties-special.html:13
    [[IsGenerator]]: true
    [[Scopes]]: Scopes[1]
